backend/requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
Django==4.2.2
django-cors-headers==4.1.0
django-filter==23.2
djangorestframework==3.14.0
docopt==0.6.2
google-api-core==2.14.0
google-auth==2.25.1
google-cloud-core==2.3.3
google-cloud-storage==2.13.0
google-crc32c==1.5.0
google-resumable-media==2.6.0
googleapis-common-protos==1.61.0
gunicorn==21.2.0
idna==3.6
importlib-metadata==6.7.0
Markdown==3.4.3
packaging==23.2
Pillow==10.1.0
pipreqs==0.4.13
protobuf==4.25.1
psycopg2==2.9.6
pyasn1==0.5.1
pyasn1-modules==0.3.0
python-dotenv==1.0.1
pytz==2023.3
requests==2.31.0
rsa==4.9
sqlparse==0.4.4
typing_extensions==4.6.3
urllib3==2.1.0
whitenoise==6.6.0
yarg==0.1.9
zipp==3.15.0
